bin/util.py: Factor out reporter class. Define __all__
